add facility controller

+++ src/main/java/com/github/fredO1211/booking/repository/GuestRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GuestRepository extends JpaRepository<Guest, Integer> {
+public interface GuestRepository extends JpaRepository<Guest, Long> {
     @Query(nativeQuery = true, value = "SELECT * FROM GUESTS where name like %:name% and phone_number like %:phoneNumber% and email like %:email%")
     List<Guest> findGuestByProps(String name, String phoneNumber, String email);
 }
